requirements.txt
slack_bolt
google-adk
python-dotenv
google-genai
google-cloud-aiplatform
google-auth
pyyaml
google-cloud-discoveryengine
